// app/themes/types/typography/gTypography.ts
export type Font = {
  fontWeight: 'bold' | 'normal' | 'bolder' | 'lighter' | number;
  fontSize: string;
  lineHeight: string | number;
};

export type TypoGraphy = {
  type100: Font;
  type101: Font;
  type102: Font;
  type120: Font;
  type121: Font;
  type122: Font;
  type140: Font;
  type141: Font;
  type142: Font;
  type160: Font;
  type161: Font;
  type162: Font;
  type200: Font;
  type201: Font;
  type202: Font;
  type240: Font;
  type241: Font;
  type242: Font;
  type320: Font;
  type321: Font;
  type322: Font;
  type360: Font;
  type361: Font;
  type362: Font;
  spacing50: number;
  spacing75: number;
  spacing100: number;
  spacing200: number;
  spacing300: number;
  spacing400: number;
  spacing500: number;
};